migrate file-read tool to schema validation
- Replace old FileReadTool with new schema-based implementation
- Delete old implementation completely
- Update executor to support both old and new tool interfaces during migration
- Update tests to work with new Tool interface
- Add annotations property to new Tool base class for compatibility
- Maintain all existing functionality while reducing validation code by ~70%
- Improve error messages for better AI guidance
- All integration tests passing
