+++ src/tools/__tests__/file-read.test.ts
@@ -5,7 +5,6 @@ import { describe, it, expect, beforeAll, afterAll } from 'vitest';
 import { writeFile, mkdir, rm } from 'fs/promises';
 import { join } from 'path';
 import { FileReadTool } from '../implementations/file-read.js';
-import { createTestToolCall } from './test-utils.js';
 
 describe('FileReadTool', () => {
   const tool = new FileReadTool();
@@ -25,26 +24,25 @@ describe('FileReadTool', () => {
   describe('tool metadata', () => {
     it('should have correct name and description', () => {
       expect(tool.name).toBe('file_read');
-      expect(tool.description).toBe('Read file contents with optional line range');
-      expect(tool.annotations?.readOnlyHint).toBe(true);
+      expect(tool.description).toBe('Read file contents with optional line range support');
+      // New tool doesn't have annotations property - that's old interface
     });
 
     it('should have correct input schema', () => {
-      expect(tool.inputSchema).toEqual({
-        type: 'object',
-        properties: {
-          path: { type: 'string', description: 'File path to read' },
-          startLine: { type: 'number', description: 'Starting line number (1-based, optional)' },
-          endLine: { type: 'number', description: 'Ending line number (1-based, optional)' },
-        },
-        required: ['path'],
-      });
+      const schema = tool.inputSchema;
+      expect(schema.type).toBe('object');
+      expect(schema.properties.path).toBeDefined();
+      expect(schema.properties.startLine).toBeDefined();
+      expect(schema.properties.endLine).toBeDefined();
+      expect(schema.required).toContain('path');
+      expect(schema.required).not.toContain('startLine');
+      expect(schema.required).not.toContain('endLine');
     });
   });
 
   describe('file reading', () => {
     it('should read entire file when no range specified', async () => {
-      const result = await tool.executeTool(createTestToolCall('file_read', { path: testFile }));
+      const result = await tool.execute({ path: testFile });
 
       expect(result.isError).toBe(false);
       expect(result.content).toHaveLength(1);
@@ -53,37 +51,31 @@ describe('FileReadTool', () => {
     });
 
     it('should read specific line range', async () => {
-      const result = await tool.executeTool(
-        createTestToolCall('file_read', {
-          path: testFile,
-          startLine: 2,
-          endLine: 4,
-        })
-      );
+      const result = await tool.execute({
+        path: testFile,
+        startLine: 2,
+        endLine: 4,
+      });
 
       expect(result.isError).toBe(false);
       expect(result.content[0].text).toBe('Line 2\nLine 3\nLine 4');
     });
 
     it('should read from start line to end of file', async () => {
-      const result = await tool.executeTool(
-        createTestToolCall('file_read', {
-          path: testFile,
-          startLine: 3,
-        })
-      );
+      const result = await tool.execute({
+        path: testFile,
+        startLine: 3,
+      });
 
       expect(result.isError).toBe(false);
       expect(result.content[0].text).toBe('Line 3\nLine 4\nLine 5');
     });
 
     it('should read from beginning to end line', async () => {
-      const result = await tool.executeTool(
-        createTestToolCall('file_read', {
-          path: testFile,
-          endLine: 2,
-        })
-      );
+      const result = await tool.execute({
+        path: testFile,
+        endLine: 2,
+      });
 
       expect(result.isError).toBe(false);
       expect(result.content[0].text).toBe('Line 1\nLine 2');
@@ -92,54 +84,45 @@ describe('FileReadTool', () => {
 
   describe('error handling', () => {
     it('should handle missing path parameter', async () => {
-      const result = await tool.executeTool(createTestToolCall('file_read', {}));
+      const result = await tool.execute({});
 
       expect(result.isError).toBe(true);
-      expect(result.content[0].text).toBe(
-        "Parameter 'path' must be string. Provide a valid string value. Parameter is required"
-      );
+      expect(result.content[0].text).toContain('Validation failed');
+      expect(result.content[0].text).toContain('path: Required');
     });
 
     it('should handle empty path parameter', async () => {
-      const result = await tool.executeTool(createTestToolCall('file_read', { path: '' }));
+      const result = await tool.execute({ path: '' });
 
       expect(result.isError).toBe(true);
-      expect(result.content[0].text).toBe(
-        "Parameter 'path' must be non-empty string. Provide a valid non-empty string value. Parameter cannot be empty"
-      );
+      expect(result.content[0].text).toContain('Validation failed');
+      expect(result.content[0].text).toContain('File path cannot be empty');
     });
 
     it('should handle non-existent file', async () => {
-      const result = await tool.executeTool(
-        createTestToolCall('file_read', { path: '/non/existent/file.txt' })
-      );
+      const result = await tool.execute({ path: '/non/existent/file.txt' });
 
       expect(result.isError).toBe(true);
-      expect(result.content[0].text).toContain('File reading failed');
+      expect(result.content[0].text).toContain('File not found');
     });
 
     it('should handle start line beyond file length', async () => {
-      const result = await tool.executeTool(
-        createTestToolCall('file_read', {
-          path: testFile,
-          startLine: 10,
-        })
-      );
+      const result = await tool.execute({
+        path: testFile,
+        startLine: 10,
+      });
 
       expect(result.isError).toBe(true);
-      expect(result.content[0].text).toBe(
-        'File reading failed. Check the file path and line numbers, then try again. Unknown error occurred'
-      );
+      expect(result.content[0].text).toContain('Line 10 exceeds file length');
+      expect(result.content[0].text).toContain('Use a line number between 1 and 5');
     });
 
     it('should handle end line beyond file length gracefully', async () => {
-      const result = await tool.executeTool(
-        createTestToolCall('file_read', {
-          path: testFile,
-          startLine: 3,
-          endLine: 10,
-        })
-      );
+      const result = await tool.execute({
+        path: testFile,
+        startLine: 3,
+        endLine: 10,
+      });
 
       expect(result.isError).toBe(false);
       expect(result.content[0].text).toBe('Line 3\nLine 4\nLine 5');
@@ -151,11 +134,9 @@ describe('FileReadTool', () => {
       const largeFile = join(testDir, 'large.txt');
       await writeFile(largeFile, largeContent);
 
-      const result = await tool.executeTool(
-        createTestToolCall('file_read', {
-          path: largeFile,
-        })
-      );
+      const result = await tool.execute({
+        path: largeFile,
+      });
 
       expect(result.isError).toBe(true);
       expect(result.content[0].text).toContain('File is too large');
@@ -172,13 +153,11 @@ describe('FileReadTool', () => {
       const largeFile = join(testDir, 'large-lines.txt');
       await writeFile(largeFile, largeContent);
 
-      const result = await tool.executeTool(
-        createTestToolCall('file_read', {
-          path: largeFile,
-          startLine: 500,
-          endLine: 505,
-        })
-      );
+      const result = await tool.execute({
+        path: largeFile,
+        startLine: 500,
+        endLine: 505,
+      });
 
       expect(result.isError).toBe(false);
       expect(result.content[0].text).toContain('Line 500');
@@ -186,13 +165,11 @@ describe('FileReadTool', () => {
     });
 
     it('should reject ranged reads larger than 100 lines', async () => {
-      const result = await tool.executeTool(
-        createTestToolCall('file_read', {
-          path: testFile,
-          startLine: 1,
-          endLine: 101, // 101 lines = too large
-        })
-      );
+      const result = await tool.execute({
+        path: testFile,
+        startLine: 1,
+        endLine: 101, // 101 lines = too large
+      });
 
       expect(result.isError).toBe(true);
       expect(result.content[0].text).toContain('Range too large (101 lines)');
@@ -205,7 +182,7 @@ describe('FileReadTool', () => {
       const emptyFile = join(testDir, 'empty.txt');
       await writeFile(emptyFile, '');
 
-      const result = await tool.executeTool(createTestToolCall('file_read', { path: emptyFile }));
+      const result = await tool.execute({ path: emptyFile });
 
       expect(result.isError).toBe(false);
       expect(result.content[0].text).toBe('');
@@ -215,9 +192,7 @@ describe('FileReadTool', () => {
       const singleLineFile = join(testDir, 'single.txt');
       await writeFile(singleLineFile, 'Only line');
 
-      const result = await tool.executeTool(
-        createTestToolCall('file_read', { path: singleLineFile })
-      );
+      const result = await tool.execute({ path: singleLineFile });
 
       expect(result.isError).toBe(false);
       expect(result.content[0].text).toBe('Only line');
