+++ src/tools/executor.ts
@@ -2,6 +2,7 @@
 // ABOUTME: Handles tool registration, approval checks, and safe execution with simple configuration
 
 import { ToolResult, ToolContext, Tool, ToolCall, createErrorResult } from './types.js';
+import { Tool as NewTool } from './tool.js';
 import { ApprovalCallback, ApprovalDecision } from './approval-types.js';
 import { BashTool } from './implementations/bash.js';
 import { FileReadTool } from './implementations/file-read.js';
@@ -15,15 +16,20 @@ import { TaskAddTool, TaskListTool, TaskCompleteTool } from './implementations/t
 import { DelegateTool } from './implementations/delegate.js';
 import { UrlFetchTool } from './implementations/url-fetch.js';
 
+// Type guard to check if a tool is the new schema-based Tool class
+function isNewTool(tool: unknown): tool is NewTool {
+  return tool instanceof NewTool;
+}
+
 export class ToolExecutor {
-  private tools = new Map<string, Tool>();
+  private tools = new Map<string, Tool | NewTool>();
   private approvalCallback?: ApprovalCallback;
 
-  registerTool(name: string, tool: Tool): void {
+  registerTool(name: string, tool: Tool | NewTool): void {
     this.tools.set(name, tool);
   }
 
-  registerTools(tools: Tool[]): void {
+  registerTools(tools: (Tool | NewTool)[]): void {
     for (const tool of tools) {
       this.tools.set(tool.name, tool);
     }
@@ -33,7 +39,7 @@ export class ToolExecutor {
     this.approvalCallback = callback;
   }
 
-  getTool(toolName: string): Tool | undefined {
+  getTool(toolName: string): Tool | NewTool | undefined {
     return this.tools.get(toolName);
   }
 
@@ -41,7 +47,7 @@ export class ToolExecutor {
     return Array.from(this.tools.keys());
   }
 
-  getAllTools(): Tool[] {
+  getAllTools(): (Tool | NewTool)[] {
     return Array.from(this.tools.values());
   }
 
@@ -52,7 +58,7 @@ export class ToolExecutor {
   registerAllAvailableTools(): void {
     const tools = [
       new BashTool(),
-      new FileReadTool(),
+      new FileReadTool(), // Schema-based file read tool
       new FileWriteTool(),
       new FileEditTool(),
       new FileInsertTool(),
@@ -97,7 +103,16 @@ export class ToolExecutor {
 
     // 3. Execute the tool
     try {
-      const result = await tool.executeTool(call, context);
+      let result: ToolResult;
+
+      if (isNewTool(tool)) {
+        // New schema-based tool
+        result = await tool.execute(call.arguments, context);
+      } else {
+        // Old tool interface
+        result = await tool.executeTool(call, context);
+      }
+
       // Ensure the result has the call ID if it wasn't set by the tool
       if (!result.id && call.id) {
         result.id = call.id;
