+++ src/tools/implementations/file-read.ts
@@ -1,142 +1,191 @@
-// ABOUTME: File reading tool with optional line range support
-// ABOUTME: Safe file access for code inspection and analysis
+// ABOUTME: File reading tool with schema-based validation and line range support
+// ABOUTME: Safe file access with AI-optimized error messages and misspelling suggestions
 
-import { ToolCall, ToolResult, ToolContext } from '../types.js';
-import { BaseTool, ValidationError } from '../base-tool.js';
-import { stat } from 'fs/promises';
-import { TOOL_LIMITS } from '../constants.js';
+import { z } from 'zod';
+import { readFile, stat } from 'fs/promises';
+import { Tool } from '../tool.js';
+import { FilePath, LineNumber } from '../schemas/common.js';
+import type { ToolResult, ToolContext } from '../types.js';
+import { findSimilarPaths } from '../utils/file-suggestions.js';
 
-export class FileReadTool extends BaseTool {
+const MAX_FILE_SIZE = 32 * 1024; // 32KB limit for whole file reads
+const MAX_RANGE_SIZE = 100; // Maximum lines in a ranged read
+
+const fileReadSchema = z
+  .object({
+    path: FilePath,
+    startLine: LineNumber.optional(),
+    endLine: LineNumber.optional(),
+  })
+  .refine(
+    (data) => {
+      if (data.startLine && data.endLine) {
+        return data.endLine >= data.startLine;
+      }
+      return true;
+    },
+    {
+      message: 'endLine must be >= startLine',
+      path: ['endLine'],
+    }
+  );
+
+export class FileReadTool extends Tool {
   name = 'file_read';
-  description = 'Read file contents with optional line range';
+  description = 'Read file contents with optional line range support';
+  schema = fileReadSchema;
   annotations = {
-    title: 'File Reader',
     readOnlyHint: true,
     idempotentHint: true,
   };
-  inputSchema = {
-    type: 'object' as const,
-    properties: {
-      path: { type: 'string', description: 'File path to read' },
-      startLine: { type: 'number', description: 'Starting line number (1-based, optional)' },
-      endLine: { type: 'number', description: 'Ending line number (1-based, optional)' },
-    },
-    required: ['path'],
-  };
 
-  async executeTool(call: ToolCall, _context?: ToolContext): Promise<ToolResult> {
+  protected async executeValidated(
+    args: z.infer<typeof fileReadSchema>,
+    _context?: ToolContext
+  ): Promise<ToolResult> {
     try {
-      const path = this.validateNonEmptyStringParam(call.arguments.path, 'path', call.id);
-      const startLine = this.validateOptionalParam(
-        call.arguments.startLine,
-        'startLine',
-        (value) => this.validateNumberParam(value, 'startLine', call.id, { min: 1, integer: true }),
-        call.id
-      );
-      const endLine = this.validateOptionalParam(
-        call.arguments.endLine,
-        'endLine',
-        (value) => this.validateNumberParam(value, 'endLine', call.id, { min: 1, integer: true }),
-        call.id
-      );
-
-      // Validate startLine <= endLine if both are provided
-      if (startLine !== undefined && endLine !== undefined && startLine > endLine) {
-        return this.createStructuredError(
-          `Start line ${startLine} is greater than end line ${endLine}`,
-          'Ensure startLine <= endLine',
-          'Invalid line range specified',
-          call.id
-        );
-      }
-
-      // Validate range size limit
-      if (startLine !== undefined && endLine !== undefined) {
-        const rangeSize = endLine - startLine + 1;
-        if (rangeSize > TOOL_LIMITS.MAX_RANGE_SIZE) {
-          const error = this.createStructuredError(
-            `Range too large (${rangeSize} lines)`,
-            `Use smaller ranges (max ${TOOL_LIMITS.MAX_RANGE_SIZE} lines per read)`,
-            'Large ranges can cause context overflow and performance issues',
-            call.id
-          );
-          throw new ValidationError(error);
+      // Check file size before reading (unless using line range)
+      if (!args.startLine && !args.endLine) {
+        const sizeError = await this.validateFileSizeForWholeRead(args.path);
+        if (sizeError) {
+          return {
+            content: [{ type: 'text', text: sizeError }],
+            isError: true,
+          };
         }
       }
 
-      // Check file size before reading whole file (unless using range)
-      if (startLine === undefined && endLine === undefined) {
-        await this.validateFileSizeForWholeRead(path, call.id);
+      // Validate range size if both start and end are specified
+      if (args.startLine && args.endLine) {
+        const rangeSize = args.endLine - args.startLine + 1;
+        if (rangeSize > MAX_RANGE_SIZE) {
+          return {
+            content: [
+              {
+                type: 'text',
+                text: `Range too large (${rangeSize} lines). Use smaller ranges (max ${MAX_RANGE_SIZE} lines per read) to avoid context overflow and performance issues.`,
+              },
+            ],
+            isError: true,
+          };
+        }
       }
 
-      // Read file and split into lines
-      const content = await this.readFileWithContext(path, call.id);
+      // Read file content
+      const content = await readFile(args.path, 'utf-8');
       const lines = content.split('\n');
 
-      let resultLines = lines;
-      if (startLine !== undefined || endLine !== undefined) {
-        const start = Math.max(0, (startLine ?? 1) - 1);
-        const end = endLine !== undefined ? Math.min(lines.length, endLine) : lines.length;
-
-        // Validate line numbers against file content
-        if (startLine !== undefined) {
-          this.validateLineNumber(startLine, content, 'startLine', call.id);
-        }
-        // For endLine, we allow it to exceed file length (will just return available lines)
-
-        resultLines = lines.slice(start, end);
+      // Validate line numbers against actual file content
+      if (args.startLine && args.startLine > lines.length) {
+        return {
+          content: [
+            {
+              type: 'text',
+              text: `Line ${args.startLine} exceeds file length (${lines.length} lines). Use a line number between 1 and ${lines.length}.`,
+            },
+          ],
+          isError: true,
+        };
       }
 
-      const resultContent = resultLines.join('\n');
-      const totalLines = this.countLines(content);
-      const metadata = {
-        totalLines,
-        linesReturned: resultLines.length,
-        fileSize: this.formatFileSize(content.length),
-      };
+      // Apply line range if specified
+      let resultContent = content;
+      let linesReturned = lines.length;
 
-      return this.createSuccessWithMetadata(
-        [
+      if (args.startLine || args.endLine) {
+        const start = Math.max(0, (args.startLine || 1) - 1);
+        const end =
+          args.endLine !== undefined ? Math.min(lines.length, args.endLine) : lines.length;
+
+        const resultLines = lines.slice(start, end);
+        resultContent = resultLines.join('\n');
+        linesReturned = resultLines.length;
+      }
+
+      return {
+        content: [
           {
             type: 'text',
             text: resultContent,
           },
         ],
-        metadata,
-        call.id
-      );
-    } catch (error) {
-      if (error instanceof ValidationError) {
-        return error.toolResult;
+        isError: false,
+        metadata: {
+          totalLines: lines.length,
+          linesReturned,
+          fileSize: this.formatFileSize(content.length),
+        },
+      };
+    } catch (error: any) {
+      // Handle file not found with helpful suggestions
+      if (error.code === 'ENOENT') {
+        const suggestions = await findSimilarPaths(args.path);
+        const suggestionText =
+          suggestions.length > 0 ? `\n\nSimilar files: ${suggestions.join(', ')}` : '';
+
+        return {
+          content: [
+            {
+              type: 'text',
+              text: `File not found: ${args.path}${suggestionText}`,
+            },
+          ],
+          isError: true,
+        };
       }
 
-      return this.createStructuredError(
-        'File reading failed',
-        'Check the file path and line numbers, then try again',
-        error instanceof Error ? error.message : 'Unknown error occurred',
-        call.id
-      );
+      // Handle permission errors
+      if (error.code === 'EACCES') {
+        return {
+          content: [
+            {
+              type: 'text',
+              text: `Permission denied accessing file: ${args.path}. Check file permissions and try again.`,
+            },
+          ],
+          isError: true,
+        };
+      }
+
+      // Handle other file system errors
+      if (error.code) {
+        return {
+          content: [
+            {
+              type: 'text',
+              text: `File system error (${error.code}): ${error.message}. Check the file path and try again.`,
+            },
+          ],
+          isError: true,
+        };
+      }
+
+      // Re-throw unexpected errors
+      throw error;
     }
   }
 
-  private async validateFileSizeForWholeRead(filePath: string, callId?: string): Promise<void> {
+  private async validateFileSizeForWholeRead(filePath: string): Promise<string | null> {
     try {
       const fileStats = await stat(filePath);
-      if (fileStats.size > TOOL_LIMITS.MAX_WHOLE_FILE_SIZE) {
+      if (fileStats.size > MAX_FILE_SIZE) {
         const fileSizeFormatted = this.formatFileSize(fileStats.size);
-        const error = this.createStructuredError(
-          `File is too large (${fileSizeFormatted}) for whole-file read`,
-          `Use startLine and endLine parameters for ranged reads (e.g., startLine: 1, endLine: 100)`,
-          `File size limit is ${this.formatFileSize(TOOL_LIMITS.MAX_WHOLE_FILE_SIZE)} for whole-file reads`,
-          callId
-        );
-        throw new ValidationError(error);
+        return `File is too large (${fileSizeFormatted}) for whole-file read. Use startLine and endLine parameters for ranged reads (e.g., startLine: 1, endLine: 100). File size limit is ${this.formatFileSize(
+          MAX_FILE_SIZE
+        )} for whole-file reads.`;
       }
-    } catch (error) {
-      if (error instanceof ValidationError) {
-        throw error;
-      }
-      // If we can't stat the file, let readFileWithContext handle the error
+      return null;
+    } catch {
+      // If we can't stat the file, let the main read operation handle the error
+      return null;
     }
+  }
+
+  private formatFileSize(bytes: number): string {
+    if (bytes === 0) return '0 bytes';
+    const k = 1024;
+    const sizes = ['bytes', 'KB', 'MB', 'GB'];
+    const i = Math.floor(Math.log(bytes) / Math.log(k));
+    return `${parseFloat((bytes / Math.pow(k, i)).toFixed(1))} ${sizes[i]}`;
   }
 }
